3340 removing service when deleting a doc element
