+++ src/components/documentBuilder/edit/RemoveElement.tsx
@@ -15,13 +15,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { useField } from "formik";
+import { getIn, useFormikContext } from "formik";
 import React from "react";
 import { Button } from "react-bootstrap";
-import { DocumentElement } from "@/components/documentBuilder/documentBuilderTypes";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrash } from "@fortawesome/free-solid-svg-icons";
 import getElementCollectionName from "./getElementCollectionName";
+import { produceExcludeUnusedDependencies } from "@/components/fields/schemaFields/serviceFieldUtils";
+import { FormState } from "@/pageEditor/pageEditorTypes";
+import { produce } from "immer";
 
 type RemoveElementProps = {
   elementName: string;
@@ -32,22 +34,28 @@ const RemoveElement: React.FC<RemoveElementProps> = ({
   elementName,
   resetActiveElement,
 }) => {
+  const { values: formState, setValues: setFormState } =
+    useFormikContext<FormState>();
+
   // Gives the name of the elements's collection
   // In case of a list item element point to the collection of the list element,
   // i.e. removing the item of the list will actually remove the list itself.
   const { collectionName, elementIndex } =
     getElementCollectionName(elementName);
 
-  const [{ value: elementsCollection }, , { setValue }] =
-    useField<DocumentElement[]>(collectionName);
-
   const removeElement = () => {
-    const newElementsCollection = [...elementsCollection];
-    newElementsCollection.splice(Number(elementIndex), 1);
-
     resetActiveElement();
 
-    setValue(newElementsCollection);
+    // Remove the element from the form state
+    let nextState = produce(formState, (draft) => {
+      const elementsCollection = getIn(draft, collectionName);
+      elementsCollection.splice(Number(elementIndex), 1);
+    });
+
+    // If the element used a service, remove the service link as well
+    nextState = produceExcludeUnusedDependencies(nextState);
+
+    setFormState(nextState);
   };
 
   return (
